@@ -12,14 +12,14 @@ import sys
 
 def version():
     loader = importlib.machinery.SourceFileLoader(
-        "hiredis.version", "hiredis/version.py")
+        "libvalkey.version", "libvalkey/version.py")
     module = loader.load_module()
     return module.__version__
 
 
 def get_sources():
-    hiredis_sources = ("alloc", "async", "hiredis", "net", "read", "sds", "sockcompat")
-    return sorted(glob.glob("src/*.c") + ["vendor/hiredis/%s.c" % src for src in hiredis_sources])
+    libvalkey_sources = ("alloc", "async", "dict", "net", "read", "sds", "sockcompat", "valkey")
+    return sorted(glob.glob("src/*.c") + ["vendor/libvalkey/src/%s.c" % src for src in libvalkey_sources])
 
 
 def get_linker_args():
@@ -43,31 +43,31 @@ def get_libraries():
         return []
 
 
-ext = Extension("hiredis.hiredis",
+ext = Extension("libvalkey.libvalkey",
                 sources=get_sources(),
                 extra_compile_args=get_compiler_args(),
                 extra_link_args=get_linker_args(),
                 libraries=get_libraries(),
-                include_dirs=["vendor"])
+                include_dirs=["vendor/libvalkey/include", "vendor/libvalkey/include/valkey"])
 
 setup(
-    name="hiredis",
+    name="libvalkey",
     version=version(),
-    description="Python wrapper for hiredis",
+    description="Python wrapper for libvalkey",
     long_description=io.open('README.md', 'rt', encoding='utf-8').read(),
     long_description_content_type='text/markdown',
-    url="https://github.com/redis/hiredis-py",
-    author="Jan-Erik Rediger, Pieter Noordhuis",
-    author_email="janerik@fnordig.de, pcnoordhuis@gmail.com",
-    keywords=["Redis"],
+    url="https://github.com/valkey-io/libvalkey-py",
+    author="libvalkey-py authors",
+    author_email="libvalkey-py@lists.valkey.io",
+    keywords=["Valkey"],
     license="MIT",
-    packages=["hiredis"],
-    package_data={"hiredis": ["hiredis.pyi", "py.typed"]},
+    packages=["libvalkey"],
+    package_data={"libvalkey": ["libvalkey.pyi", "py.typed"]},
     ext_modules=[ext],
-    python_requires=">=3.7",
+    python_requires=">=3.8",
     project_urls={
-        "Changes": "https://github.com/redis/hiredis-py/releases",
-        "Issue tracker": "https://github.com/redis/hiredis-py/issues",
+        "Changes": "https://github.com/valkey-io/libvalkey-py/releases",
+        "Issue tracker": "https://github.com/valkey-io/libvalkey-py/issues",
     },
     classifiers=[
         'Development Status :: 5 - Production/Stable',
@@ -78,7 +78,6 @@ setup(
         'Programming Language :: C',
         'Programming Language :: Python :: 3',
         'Programming Language :: Python :: 3 :: Only',
-        'Programming Language :: Python :: 3.7',
         'Programming Language :: Python :: 3.8',
         'Programming Language :: Python :: 3.9',
         'Programming Language :: Python :: 3.10',
